requirements.txt
opencv-python==4.8.1.78
Pillow==9.2.0
face_recognition==1.3.0
cmake==3.17.2
dlib-bin
opencv-python==4.6.0.66
Pillow==9.2.0
face_recognition==1.3.0

